♻️ [Refactor]: createstudy component 이동
