+++ src/features/create-study/components/ContentAdderBox.tsx
