+++ src/features/create-study/components/ContentInputBox.tsx
