+++ src/features/create-study/components/LabelEditor.tsx
