+++ src/features/create-study/components/StudyBasicInfo.tsx
