+++ src/features/create-study/components/StudyBenefits.tsx
