+++ src/features/create-study/components/StudyCurriculum.tsx
