+++ src/features/create-study/components/StudyDescription.tsx
