+++ src/features/create-study/components/StudyMethod.tsx
