+++ src/features/create-study/components/StudyRules.tsx
